requirements.txt
absl-py==2.0.0
attrs==23.1.0
cffi==1.16.0
contourpy==1.1.1
cycler==0.12.1
flatbuffers==23.5.26
fonttools==4.43.1
frame==0.3.1
importlib-resources==6.1.0
Jinja2==3.1.2
keyboard==0.13.5
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
mediapipe==0.10.5
multipart==0.2.4
numpy==1.26.0
opencv-contrib-python==4.8.1.78
opencv-python==4.8.0.76
packaging==23.2
Pillow==10.0.1
protobuf==3.20.3
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
repoze.lru==0.7
Routes==2.5.1
six==1.16.0
sounddevice==0.4.6
zipp==3.17.0
pyserial==3.5